+++ packages/headless/src/api/search/date/date-format.test.ts
@@ -1,8 +1,5 @@
 import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
 import {isSearchApiDate, formatDateForSearchApi} from './date-format';
-
-dayjs.extend(utc);
 
 describe('#isSearchApiDate', () => {
   it('when the string matches the search api format, it returns true', () => {
@@ -15,7 +12,8 @@ describe('#isSearchApiDate', () => {
 });
 
 it('formatDateForSearchApi returns the correct format', () => {
-  expect(formatDateForSearchApi(dayjs(818035920000).utc())).toBe(
-    '1995/12/04@00:12:00'
+  const date = 818035920000;
+  expect(formatDateForSearchApi(dayjs(date))).toBe(
+    dayjs(date).format('YYYY/MM/DD@HH:mm:ss')
   );
 });