+++ packages/headless/src/controllers/facets/range-facet/date-facet/date-range.ts
@@ -1,5 +1,4 @@
 import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
 import customParseFormat from 'dayjs/plugin/customParseFormat';
 import {DateRangeRequest} from '../../../../features/facets/range-facets/date-facet-set/interfaces/request';
 import {FacetValueState} from '../../../../features/facets/facet-api/value';
@@ -10,8 +9,8 @@ import {
   isRelativeDateFormat,
   RelativeDate,
 } from '../../../../api/search/date/relative-date';
+import {isUndefined} from '@coveo/bueno';
 
-dayjs.extend(utc);
 dayjs.extend(customParseFormat);
 
 type AbsoluteDate = string | number | Date;
@@ -50,7 +49,7 @@ export interface DateRangeOptions {
   /**
    * If `true`, the date will be returned unshifted. If `false`, the date will be adjusted to UTC time.
    *
-   * @defaultValue `false`
+   * @deprecated No adjusments to UTC are being made. Please use the `timezone` engine configuration option instead.
    */
   useLocalTime?: boolean;
 }
@@ -62,6 +61,12 @@ export interface DateRangeOptions {
  * @returns A new `DateRangeRequest`.
  */
 export function buildDateRange(config: DateRangeOptions): DateRangeRequest {
+  if (!isUndefined(config.useLocalTime)) {
+    console.warn(
+      'The "useLocalTime" option for "buildDateRange" is deprecated. Please use the "timezone" engine configuration option instead.'
+    );
+  }
+
   const start = buildDate(config.start, config);
   const end = buildDate(config.end, config);
   const endInclusive = config.endInclusive ?? false;
@@ -76,7 +81,7 @@ export function buildDateRange(config: DateRangeOptions): DateRangeRequest {
 }
 
 function buildDate(rawDate: DateRangeInput, options: DateRangeOptions) {
-  const {dateFormat, useLocalTime} = options;
+  const {dateFormat} = options;
   if (isRelativeDate(rawDate)) {
     return serializeRelativeDate(rawDate);
   }
@@ -96,6 +101,5 @@ function buildDate(rawDate: DateRangeInput, options: DateRangeOptions) {
     );
   }
 
-  const adjusted = useLocalTime ? date : date.utc();
-  return formatDateForSearchApi(adjusted);
+  return formatDateForSearchApi(date);
 }